requirements.txt
matplotlib
numpy
PyOpenGL
PyQt5
Pillow